+++ src/GmGM/synthetic/__init__.py
@@ -2,7 +2,8 @@ from .generate_data import DatasetGenerator, PrecMatGenerator
 from .generate_data import PrecMatErdosRenyiGilbert, PrecMatAutoregressive
 from .generate_data import PrecMatOnes, PrecMatBlob, PrecMatIndependent
 from .generate_data import GroupedPrecMatGenerator
-from .generate_data import NormalDistribution, LogNormalDistribution, ZiLNDistribution
+from .generate_data import NormalDistribution, LogNormalDistribution
+from .generate_data import ZiLNDistribution, ZiLNMultinomial
 from .validation import measure_prec_recall, plot_prec_recall
 
 __all__ = [
@@ -18,5 +19,6 @@ __all__ = [
     "plot_prec_recall",
     "NormalDistribution",
     "LogNormalDistribution",
-    "ZiLNDistribution"
+    "ZiLNDistribution",
+    "ZiLNMultinomial",
 ]